[GRASPth] Fixing joint map creation, velocities setting procedures and skin port bufferisation (#1).

+++ include/iCub/tactileGrasp/GraspThread.h
@@ -49,7 +49,7 @@ namespace iCub {
         };
 
         /** Typedef representing the mapping of each finger into the controllable joints it contains. */
-        typedef std::vector<std::vector<double> > FingerJointMap;
+        typedef std::vector<std::vector<int> > FingerJointMap;
 
         class GraspThread : public yarp::os::RateThread {
             private:
